"Removed PurchasedStory model and related code, added purchase history route and controller, updated CartController to check for authentication, removed orders and order_chapters tables, updated story show blade to display purchase options, and modified routes in web.php"
